+++ R/theme.R
@@ -47,6 +47,7 @@
 #' for a grouped forest plot.
 #' @param legend_gp \code{gpar} graphical parameters of legend, see \code{\link[grid]{gpar}}.
 #' @param legend_ncol integer; the number of columns, see \code{\link[grid]{legendGrob}}.
+#' @param legend_byrow logical indicating whether rows of the legend are filled first, see \code{\link[grid]{legendGrob}}.
 #' @param xaxis_gp \code{gpar} graphical parameters of x-axis, see \code{\link[grid]{gpar}}.
 #' @param refline_gp \code{gpar} graphical parameters of reference line, see \code{\link[grid]{gpar}}.
 #' @param vertline_lwd Line width for extra vertical line. A vector can be provided
@@ -97,6 +98,7 @@ forest_theme <- function(base_size = 12,
                          legend_value = "",
                          legend_gp = gpar(fontsize = base_size, fontfamily = base_family, cex = 1),
                          legend_ncol = 1,
+                         legend_byrow = TRUE,
                          # X-axis
                          xaxis_gp = gpar(fontsize = base_size, fontfamily = base_family, lwd = 0.6, cex = 1),
                          # Reference line
@@ -261,7 +263,8 @@ forest_theme <- function(base_size = 12,
                       name = legend_name,
                       position = legend_position,
                       label = legend_value,
-                      ncol = legend_ncol)
+                      ncol = legend_ncol,
+                      byrow = legend_byrow)
 
     # Title
     # For backward compatability