+++ NEWS.md
@@ -1,6 +1,7 @@
 # forestploter 1.1.3
 
 * Extend vertical line to the top and the bottom.
+* Allow multiple lines of legend with `legend_ncol` in `forest_theme`.
 * Fix error in vignette.